duplication fixes
