+++ package.json
@@ -86,7 +86,8 @@
     "slugify": "^1.4.6",
     "swr": "^0.3.11",
     "use-dark-mode": "^2.3.1",
-    "web3": "^1.3.1",
+    "web3": "^1.3.4",
+    "web3modal": "^1.9.3",
     "yup": "^0.32.6"
   },
   "devDependencies": {
